+++ Pandora/Pattern/Functor/Determinable.hs
@@ -1,7 +1,0 @@
-module Pandora.Pattern.Functor.Determinable where
-
-import Pandora.Pattern.Functor.Contravariant (Contravariant)
-
-class Contravariant t => Determinable t where
-	{-# MINIMAL determine #-}
-	determine :: t a